[index] Allow index units to depend on other units without specifying a file dependency
We now consider a negative dependency index to mean "no file
dependency". This is useful if you want to write a unit file that just
depends on other units.

+++ include/indexstore/indexstore.h
@@ -354,7 +354,7 @@ indexstore_unit_dependency_get_filepath(indexstore_unit_dependency_t);
 INDEXSTORE_PUBLIC indexstore_string_ref_t
 indexstore_unit_dependency_get_name(indexstore_unit_dependency_t);
 
-INDEXSTORE_PUBLIC size_t
+INDEXSTORE_PUBLIC int
 indexstore_unit_dependency_get_index(indexstore_unit_dependency_t);
 
 #if INDEXSTORE_HAS_BLOCKS
